add #11 GeoAPIの開発用UIをコンポーネント化
